working on product slice
