+++ client/app/store.js
@@ -1,9 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 import logger from 'redux-logger';
 import authReducer from '../features/auth/authSlice';
+import productReducer from './productSlice';
 
 const store = configureStore({
-  reducer: { auth: authReducer },
+  reducer: {
+    auth: authReducer,
+    product: productReducer,
+   },
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
 });
