+++ client/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import store from './app/store.js';
-import App from './app/App';
+import App from './app/App.js';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 const root = createRoot(document.getElementById('app'));
